predicter/requirements.txt
xgboost==2.0.3
lightgbm
numpy
pandas
psycopg2-binary
